fix: udpate langchain
