+++ src/retrievals/models/embedding_auto.py
@@ -228,15 +228,6 @@ class AutoModelForEmbedding(nn.Module):
         else:
             raise ValueError(f'Input type: {type(inputs)}')
 
-    def embed_documents(self, texts: List[str]) -> List[List[float]]:
-        """Compute doc embeddings using a HuggingFace transformer model."""
-        embeddings = self.encode(texts, **self.encode_kwargs)
-        return embeddings.tolist()
-
-    def embed_query(self, text: str) -> List[float]:
-        """Compute query embeddings using a HuggingFace transformer model."""
-        return self.embed_documents([text])[0]
-
     def encode_from_loader(
         self,
         loader: DataLoader,
